Version 1.4.2 fixed bugs, cleaned up CSS, added paragraph-start class

+++ trunk/script/shared.pl
@@ -410,9 +410,9 @@ sub mpgPages2Chapter($$$$) {
       
       if ($mpgIsMultiPage{"$book-$ch"} eq "false") {next;}
       
-      for ($pg=1; $pg<=$lastPage{$book."-".$ch}; $pg++) {
-        if (!$pages{"$book-$ch-$pg"}) {next;}
-        $thispage = "$dir/$book/$prefix$book-$ch-$pg.mpg";
+      for ($pg=0; $pg<=$lastPage{$book."-".$ch}; $pg++) {
+        my $thispage = "$dir/$book/$prefix$book-$ch-$pg.mpg";
+        if (!-e $thispage) {next;}
         if (!(-e "$dir/videotmp/chapter.mpg")) {`cp $thispage $dir/videotmp/chapter.mpg`;}
         else {
           `cat $dir/videotmp/chapter.mpg $thispage > $dir/videotmp/tmp.mpg`;
@@ -424,7 +424,8 @@ sub mpgPages2Chapter($$$$) {
       if (-e "$dir/videotmp/chapter.mpg") {
 		  if ($postFlag == 1) {
 			`ffmpeg -i "$dir/videotmp/chapter.mpg" -vcodec copy -acodec libmp3lame -y "$dir/$book/fin-$book-$ch.mpg"`;
-			`rm "$dir/videotmp/chapter.mpg"`;
+			if (!$debug) {`rm "$dir/videotmp/chapter.mpg"`;}
+			else {`mv "$dir/videotmp/chapter.mpg" "$dir/$book/tmp-$book-$ch.mpg"`;}
 		  }
 		  else {`mv $dir/videotmp/chapter.mpg $dir/$book/fin-$book-$ch.mpg`;}
 	  }
